pinakes/scrapers/papyri_scraper.py
#!/usr/bin/env python3
"""
Papyri.info Scraper for CALLIMACHINA Protocol
Hunts for bibliographic ghosts in digitized papyrus collections
"""

import requests
import re
import yaml
from datetime import datetime
from typing import List, Dict, Any

class PapyriScraper:
    def __init__(self):
        self.base_url = "https://papyri.info"
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'CALLIMACHINA/2.0 (Alexandria Reconstruction Protocol)'
        })
    
    def search_fragments(self, author: str = None, work: str = None, 
                        collection: str = None, limit: int = 100) -> List[Dict[str, Any]]:
        """
        Search papyri.info for fragments matching criteria
        """
        fragments = []
        
        # Build search query
        query_params = []
        if author:
            query_params.append(f"author={author}")
        if work:
            query_params.append(f"title={work}")
        if collection:
            query_params.append(f"collection={collection}")
        
        query = "&".join(query_params) if query_params else ""
        search_url = f"{self.base_url}/search?{query}"
        
        try:
            response = self.session.get(search_url, timeout=30)
            if response.status_code == 200:
                # Parse results (simplified - would need actual HTML parsing in production)
                fragments = self._parse_search_results(response.text, author, work)
        except Exception as e:
            print(f"[PAPYRI SCRAPE ERROR] {e}")
        
        return fragments
    
    def _parse_search_results(self, html: str, author: str, work: str) -> List[Dict[str, Any]]:
        """
        Parse search results HTML to extract fragment metadata
        """
        fragments = []
        
        # Look for DDbDP (Duke Databank of Documentary Papyri) identifiers
        ddbdp_pattern = r'DDbDP;([^;]+);([^;]+)'
        matches = re.findall(ddbdp_pattern, html)
        
        for match in matches[:50]:  # Limit initial scrape
            collection, identifier = match
            fragment = {
                'id': f"papyri.{collection}.{identifier}",
                'source': 'papyri.info',
                'collection': collection,
                'identifier': identifier,
                'author': author,
                'work': work,
                'date_scraped': datetime.now().isoformat(),
                'confidence': 0.0,  # Will be updated by stylometric analysis
                'citation_chain': [],
                'survival_path': ['papyrus'],
                'text': '',  # Full text would be fetched separately
                'language': 'greek',  # Default, would be detected
                'fragment_type': 'literary' if author else 'documentary'
            }
            fragments.append(fragment)
        
        return fragments
    
    def get_oxyrhynchus_batch(self) -> List[Dict[str, Any]]:
        """
        Specifically hunt for new Oxyrhynchus papyri
        HIGH PRIORITY: These often contain literary fragments
        """
        print("[OXYRHYNCHUS HUNT] Scanning for new literary fragments...")
        return self.search_fragments(collection="Oxyrhynchus")
    
    def get_herculaneum_batch(self) -> List[Dict[str, Any]]:
        """
        Scan Herculaneum papyri - often Epicurean/philosophical texts
        """
        print("[HERCULANEUM HUNT] Scanning carbonized papyri...")
        return self.search_fragments(collection="Herculaneum")
    
    def save_fragments(self, fragments: List[Dict[str, Any]], filename: str = None):
        """
        Save discovered fragments to YAML file
        """
        if not filename:
            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
            filename = f"/Volumes/VIXinSSD/callimachina/pinakes/fragments/papyri_batch_{timestamp}.yml"
        
        with open(filename, 'w') as f:
            yaml.dump({
                'batch_timestamp': datetime.now().isoformat(),
                'source': 'papyri.info',
                'fragment_count': len(fragments),
                'fragments': fragments
            }, f, default_flow_style=False)
        
        print(f"[FRAGMENTS SAVED] {len(fragments)} fragments to {filename}")
        return filename

if __name__ == "__main__":
    scraper = PapyriScraper()
    
    # Hunt for high-priority targets
    print("=" * 60)
    print("CALLIMACHINA PAPYRI SCRAPER INITIALIZATION")
    print("=" * 60)
    
    # Scan for Oxyrhynchus literary fragments
    oxy_fragments = scraper.get_oxyrhynchus_batch()
    if oxy_fragments:
        scraper.save_fragments(oxy_fragments, 
                             "/Volumes/VIXinSSD/callimachina/pinakes/fragments/oxyrhynchus_hunt.yml")
    
    # Scan for Herculaneum philosophical texts
    herc_fragments = scraper.get_herculaneum_batch()
    if herc_fragments:
        scraper.save_fragments(herc_fragments,
                             "/Volumes/VIXinSSD/callimachina/pinakes/fragments/herculaneum_hunt.yml")
    
    print("[INITIAL SCAN COMPLETE] Awaiting further instructions...")
