requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Bayesian inference
pymc>=5.0.0
arviz>=0.15.0

# Network analysis
networkx>=3.0

# Natural language processing
nltk>=3.8
scikit-learn>=1.3.0

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# CLI and database (v3.1 additions)
# sqlite3 is built-in with Python >= 3.8

# Optional: For advanced NLP
# transformers>=4.30.0
# torch>=2.0.0

# Optional: For advanced database features
# sqlalchemy>=2.0.0